Move ludwig/utils/schema file to new ludwig/schema folder.
